More tests, more bug fixes.

+++ src/include/util/defs.h
@@ -25,7 +25,7 @@
 } while (0)
 
 // FIXME: find a better place for such common defs?
-#define sizeofarray(a) (sizeof(a) / sizeof(*a))
+#define sizeofarray(a) ((size_t)(sizeof(a) / sizeof(*a)))
 #define DECONST(v)  ((void *)(uintptr_t)(v))
 #define __constructor __attribute__((__constructor__))
 #define __printflike(f,a) __attribute__((__format__(__printf__,f,a)))
